src/use-cases/get-user-profile.spec.ts
import { InMemoryUsersRepository } from "@/repositories/inMemory/in-memory-user-repository";
import { hash } from "bcryptjs";
import { beforeEach, describe, expect, it } from "vitest";
import { ResourseNotFoundError } from "./errors/resouse-not-found-error";
import { GetUserProfileService } from "./get-user-profile";

let usersRepository: InMemoryUsersRepository;
let sut: GetUserProfileService;
describe("Get User Profile Use Case", () => {
  beforeEach(() => {
    usersRepository = new InMemoryUsersRepository();
    sut = new GetUserProfileService(usersRepository);
  });

  it("should be able to get user profile", async () => {
    const createdUser = await usersRepository.create({
      name: "John Doe",
      email: "johndoe@example.com",
      password_hash: await hash("123456", 6),
    });

    const { user } = await sut.execute({
      userId: createdUser.id,
    });

    expect(user.id).toEqual(expect.any(String));
    expect(user.name).toEqual("John Doe");
  });

  it("should not to be able toget user profile wrong id", async () => {
    await expect(() =>
      sut.execute({
        userId: "non-existing-id",
      }),
    ).rejects.toBeInstanceOf(ResourseNotFoundError);
  });
});
